Moving some files around to remove dependencies
