--- externalpackages/durabletaskscheduler/__init__.py
@@ -1,7 +1,0 @@
-# Copyright (c) Microsoft Corporation.
-# Licensed under the MIT License.
-
-"""Durable Task SDK for Python"""
-
-
-PACKAGE_NAME = "durabletaskscheduler"
