+++ requirements.txt
@@ -4,3 +4,4 @@ protobuf
 pytest
 pytest-cov
 azure-identity
+durabletask-azuremanaged
